<internal>: changes to ResultForm and others
in response to Result id now being long

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/ResultFormService.java
@@ -3,5 +3,5 @@ package ar.edu.itba.paw.interfaces;
 import ar.edu.itba.paw.model.ResultForm;
 
 public interface ResultFormService {
-    public Integer HandleOrderForm(ResultForm resultForm, byte[] identification, String identificationType, byte[] resultData, String resultDataType, long orderId);
+    public Long HandleOrderForm(ResultForm resultForm, byte[] identification, String identificationType, byte[] resultData, String resultDataType, long orderId);
 }
